metadata: add various missing glyphs for Vietnamese
This adds various missing glyphs for Vietnamese.
Note that some characters require multiple diacritics in different positions.

There is a hack in diacritics.json to force monocraft.py to include many
glyphs in the generation step. For example, the glyph "LATIN SMALL LETTER
A WITH CIRCUMFLEX AND DOT BELOW" contain two diacritics `circumflex` and
`dot_below`. generate_diacritics.py parses this as containing one
diacritic: `circumflex_and_dot_below`, which it will not find in
diacritics.json without this hack, and thus the glyph will not be
generated.
For a proper fix, the logic to determine which diacritics should be
considered (lines 37-38, generate_diacritics.py) should be fixed to
consult the list in characters.json instead.