generate: raise glyph ascent height
Some Vietnamese glyphs are very tall with their diacritics, which causes the top of them to sometimes be cut off in some programs. This raises the font ascent by 1 pixel size to accomodate those glyphs.
As a side effect, however, this also messes with font sizing in a lot of programs.